refactor: unify middleware directory
