--- backend/src/middleware/authMiddleware.js
@@ -1,18 +1,0 @@
-const jwt = require("jsonwebtoken");
-const User = require("../models/User");
-
-module.exports = async (req, res, next) => {
-  const token = req.headers["authorization"]?.split(" ")[1];
-  if (!token) return res.status(401).json({ message: "No token" });
-
-  try {
-    const decoded = jwt.verify(token, process.env.JWT_SECRET);
-    const user = await User.findById(decoded._id).select("-password");
-    if (!user) return res.status(401).json({ message: "User not found" });
-
-    req.user = user;
-    next();
-  } catch {
-    res.status(401).json({ message: "Invalid token" });
-  }
-};
