+++ backend/src/routes/music.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const musicController = require('../controllers/musicController');
-const auth = require('../middleware/authMiddleware');
+const auth = require('../middlewares/authMiddleware');
 
 const onlyMaster = (req, res, next) => {
   if (req.user && (req.user.role === 'master' || req.user.role === 'admin')) return next();
